Move CP solvers out of base class .h and into their own header files
